+++ magnolia/main-module/pom.xml
@@ -3,11 +3,11 @@
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <artifactId>magnolia-starter-kit</artifactId>
-    <groupId>mgnl.starter.kit</groupId>
+    <groupId>ch.openinteractive.main</groupId>
     <version>1.0-SNAPSHOT</version>
   </parent>
-  <groupId>mgnl.starter.kit</groupId>
-  <artifactId>my-module</artifactId>
+  <groupId>ch.openinteractive.main</groupId>
+  <artifactId>main-module</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
   <name>Magnolia Starter Kit Module</name>
