+++ magnolia/main-module/src/main/java/ch/openinteractive/main/StarterKit.java
@@ -1,4 +1,4 @@
-package mgnl.starter.kit;
+package ch.openinteractive.main;
 
 /**
  * This class is optional and represents the configuration for the magnolia-starter-kit module.
