+++ magnolia/main-webapp/pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>mgnl.starter.kit</groupId>
+    <groupId>ch.openinteractive.main</groupId>
     <artifactId>magnolia-starter-kit</artifactId>
     <version>1.0-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
@@ -20,7 +20,7 @@
     -->
 
     <dependency>
-      <groupId>mgnl.starter.kit</groupId>
+      <groupId>ch.openinteractive.main</groupId>
       <artifactId>main-module</artifactId>
       <version>1.0-SNAPSHOT</version>
     </dependency>
