+++ magnolia/main-webapp/src/main/webapp/WEB-INF/config/default/magnolia.properties
@@ -60,7 +60,7 @@ magnolia.connection.jcr.userId = admin
 magnolia.connection.jcr.password = admin
 
 # Set it to true if bootstrapping/update should be performed automatically
-magnolia.update.auto=true
+magnolia.update.auto=false
 
 # Location of the file containing both the private and the public keys used to verify authenticity of activation requests
 # This file is generated if not present
