+++ magnolia/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>mgnl.starter.kit</groupId>
+  <groupId>ch.openinteractive.main</groupId>
   <artifactId>magnolia-starter-kit</artifactId>
   <name>Magnolia Starter Kit (parent pom)</name>
   <version>1.0-SNAPSHOT</version>
@@ -111,7 +111,7 @@
   </repositories>
 
   <modules>
-    <module>my-webapp</module>
-    <module>my-module</module>
+    <module>main-webapp</module>
+    <module>main-module</module>
   </modules>
 </project>
